requirements.txt
Pillow==7.2.0
spidev==3.5
RPi.GPIO==0.7.0